+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/sources/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './sources'
+export { default } from './container'
